feat: protocol  AuthenticationServiceType 및 AuthenticationViewModel Action 에 checkAuthenticationState() 추가 및 파일 정리
