+++ HongikYeolgong2-iOS/Authentication/AuthenticationView.swift
@@ -1,21 +1,23 @@
-//
-//  AuthenticationView.swift
-//  HongikYeolgong2-iOS
-//
-//  Created by 변정훈 on 7/11/24.
-//
+
 
 import SwiftUI
 
 struct AuthenticationView: View {
+    @StateObject private var coordinator = SceneCoordinator()
     @StateObject var authViewModel: AuthenticationViewModel
     var body: some View {
-        switch authViewModel.authenticationState {
-        case .unauthenticated:
-            LoginView()
-                .environmentObject(authViewModel)
-        case .authenticated:
-            HomeView(viewModel: HomeViewModel())
+        VStack {
+            switch authViewModel.authenticationState {
+            case .unauthenticated:
+                LoginView()
+                    .environmentObject(authViewModel)
+            case .authenticated:
+                HomeView(viewModel: HomeViewModel())
+                    .environmentObject(coordinator)
+            }
+        }
+        .onAppear {
+            authViewModel.send(action: .checkAuthenticationState) 
         }
     }
 }
