+++ HongikYeolgong2-iOS/Scene/APP/HongikYeolgong2_iOSApp.swift
@@ -1,0 +1,31 @@
+
+
+import SwiftUI
+import Firebase
+
+@main
+struct HongikYeolgong2_iOSApp: App {
+    @UIApplicationDelegateAdaptor(AppDelegate.self) var delegate
+
+    var body: some Scene {
+        WindowGroup {
+            AuthenticationView(authViewModel: AuthenticationViewModel())
+        }
+    }
+}
+
+class AppDelegate: NSObject, UIApplicationDelegate {
+    
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
+        
+        FirebaseApp.configure()
+        registerDependency()
+        
+        return true
+    }
+    
+}
+
+func registerDependency() {
+    DIContainer.shared.register(AuthenticationService() as AuthenticationServiceType)
+}
